requirements.txt
streamlit
google-generativeai
pandas
plotly
pillow